requirements.txt
torch
torchvision
tqdm
PyYAML
tensorboard
torchinfo
matplotlib
scikit-learn
